🐞 fix(app): make widget previews run in the web previewer
- pin riverpod/flutter_riverpod/riverpod_annotation exactly (3.1.0/4.0.0)
  so the preview scaffold resolves the same versions the generated
  providers were built against
- PlatformInt64 is BigInt on web: add platformInt64() conditional-import
  helper for constructing i64 mock data, and use .toInt() where i64
  values feed int APIs (progress bars, filesize, log timestamps)
- replace Platform.pathSeparator splits with a dart:io-free
  fileBaseName() so SendPage/HomePage build in the previewer
- previewShell: key the ProviderScope per preview (riverpod forbids
  changing override sets on a live scope when the previewer swaps
  previews) and use InheritedLocaleData instead of slang's
  TranslationProvider, whose process-wide GlobalKey collides when
  several previews are mounted at once

flutter widget-preview start now loads all 12 previews with zero
exceptions.

+++ lib/common/utils.dart
@@ -122,7 +122,7 @@ void initLogger() {
         logger.log(
           levelList[event.level],
           event.msg,
-          time: DateTime.fromMillisecondsSinceEpoch(event.timeMillis),
+          time: DateTime.fromMillisecondsSinceEpoch(event.timeMillis.toInt()),
         );
       });
     }
